requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-extensions==3.2.3
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==35.2.0
flower==2.0.1
gunicorn==23.0.0
h11==0.14.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
Markdown==3.7
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
pytest==8.3.4
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomli==2.2.1
tornado==6.4.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
